Move metrics tracker from shared to plugin reporting
createMetricsTracker lived in src/shared/metrics.js but is only consumed by plugin identifier-case and project-index modules, so relocate it to src/plugin/src/reporting/metrics-tracker.js and update imports to reflect its plugin-scoped usage.
