+++ src/plugin/src/reporting/metrics-tracker.js
